# openapi.yaml
openapi: 3.1.0
info:
  title: Qarax API
  description: REST API for managing virtual machines and hypervisor hosts
  license:
    name: ''
  version: 0.1.0
paths:
...
  /networks:
    get:
      tags:
      - networks
      operationId: list
      parameters:
      - name: name
        in: query
        description: Optional name filter for list queries
        required: false
        schema:
          type:
          - string
          - 'null'
      responses:
        '200':
          description: List all networks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Network'
        '500':
          description: Internal server error
    post:
      tags:
      - networks
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewNetwork'
        required: true
      responses:
        '201':
          description: Network created successfully
          content:
            text/plain:
              schema:
                type: string
        '422':
          description: Invalid input
        '500':
          description: Internal server error
  /networks/{network_id}:
    get:
      tags:
      - networks
      operationId: get
      parameters:
      - name: network_id
        in: path
        description: Network unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Network found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Network'
        '404':
          description: Network not found
        '500':
          description: Internal server error
    delete:
      tags:
      - networks
      operationId: delete
      parameters:
      - name: network_id
        in: path
        description: Network unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Network deleted successfully
        '404':
          description: Network not found
        '500':
          description: Internal server error
  /networks/{network_id}/hosts:
    post:
      tags:
      - networks
      operationId: attach_host
      parameters:
      - name: network_id
        in: path
        description: Network unique identifier
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachHostRequest'
        required: true
      responses:
        '204':
          description: Host attached to network
        '404':
          description: Network or host not found
        '422':
          description: Failed to attach network on node
        '500':
          description: Internal server error
  /networks/{network_id}/hosts/{host_id}:
    delete:
      tags:
      - networks
      operationId: detach_host
      parameters:
      - name: network_id
        in: path
        description: Network unique identifier
        required: true
        schema:
          type: string
          format: uuid
      - name: host_id
        in: path
        description: Host unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Host detached from network
        '404':
          description: Host not found
        '422':
          description: Failed to detach network on node
        '500':
          description: Internal server error
  /networks/{network_id}/ips:
    get:
      tags:
      - networks
      operationId: list_ips
      parameters:
      - name: network_id
        in: path
        description: Network unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: List IP allocations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IpAllocation'
        '500':
          description: Internal server error
...
  /storage-pools/{pool_id}/hosts:
    post:
      tags:
      - storage-pools
      operationId: attach_host
      parameters:
      - name: pool_id
        in: path
        description: Storage pool unique identifier
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachPoolHostRequest'
        required: true
      responses:
        '204':
          description: Host attached to storage pool
        '404':
          description: Storage pool or host not found
        '500':
          description: Internal server error
  /storage-pools/{pool_id}/hosts/{host_id}:
    delete:
      tags:
      - storage-pools
      operationId: detach_host
      parameters:
      - name: pool_id
        in: path
        description: Storage pool unique identifier
        required: true
        schema:
          type: string
          format: uuid
      - name: host_id
        in: path
        description: Host unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Host detached from storage pool
        '404':
          description: Host not found
        '500':
          description: Internal server error
...
components:
  schemas:
...
    AttachHostRequest:
      type: object
      required:
      - host_id
      - bridge_name
      properties:
        bridge_name:
          type: string
        host_id:
          type: string
          format: uuid
        parent_interface:
          type:
          - string
          - 'null'
    AttachPoolHostRequest:
      type: object
      required:
      - host_id
      properties:
        host_id:
          type: string
          format: uuid
...
    IpAllocation:
      type: object
      required:
      - id
      - network_id
      - ip_address
      - allocated_at
      properties:
        allocated_at:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        ip_address:
          type: string
        network_id:
          type: string
          format: uuid
        vm_id:
          type:
          - string
          - 'null'
          format: uuid
...
    Network:
      type: object
      required:
      - id
      - name
      - subnet
      - status
      properties:
        dns:
          type:
          - string
          - 'null'
        gateway:
          type:
          - string
          - 'null'
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          $ref: '#/components/schemas/NetworkStatus'
        subnet:
          type: string
        type:
          type:
          - string
          - 'null'
        vpc_name:
          type:
          - string
          - 'null'
...
    NetworkStatus:
      type: string
      enum:
      - active
      - inactive
...
    NewNetwork:
      type: object
      required:
      - name
      - subnet
      properties:
        dns:
          type:
          - string
          - 'null'
        gateway:
          type:
          - string
          - 'null'
        name:
          type: string
        subnet:
          type: string
        type:
          type:
          - string
          - 'null'
        vpc_name:
          type:
          - string
          - 'null'
...